# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * RumoCarta API — exam prep platform for INATRO driving test
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Opcao {
  id: string;
  label: string;
  text: string;
}

export type QuestaoComRespostaMediaType = typeof QuestaoComRespostaMediaType[keyof typeof QuestaoComRespostaMediaType];


export const QuestaoComRespostaMediaType = {
  sinal: 'sinal',
  foto: 'foto',
  nenhum: 'nenhum',
} as const;

export interface QuestaoComResposta {
  id: string;
  questionText: string;
  mediaType: QuestaoComRespostaMediaType;
  /** @nullable */
  imagePath: string | null;
  categoryId: string;
  categoryName: string;
  /** @nullable */
  explanation: string | null;
  options: Opcao[];
  correctOptionId: string;
}

export interface QuestionCount {
  count: number;
}

export interface Categoria {
  id: string;
  slug: string;
  name: string;
  icon: string;
  sortOrder: number;
}

export type GetSimuladoQuestoesParams = {
/**
 * Number of questions to return
 */
count?: number;
/**
 * Filter by category ID
 */
categoryId?: string;
/**
 * Comma-separated question IDs to return (for retry-wrong-answers mode)
 */
onlyIds?: string;
};