# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: RumoCarta API — exam prep platform for INATRO driving test
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /simulado/questoes:
    get:
      operationId: getSimuladoQuestoes
      tags: [simulado]
      summary: Get random exam questions
      description: Returns N random questions with shuffled options for a simulado session
      parameters:
        - name: count
          in: query
          required: false
          schema:
            type: integer
            default: 30
          description: Number of questions to return
        - name: categoryId
          in: query
          required: false
          schema:
            type: string
          description: Filter by category ID
        - name: onlyIds
          in: query
          required: false
          schema:
            type: string
          description: Comma-separated question IDs to return (for retry-wrong-answers mode)
      responses:
        "200":
          description: List of questions with options
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/QuestaoComResposta"
...
components:
  schemas:
...
    Opcao:
      type: object
      properties:
        id:
          type: string
        label:
          type: string
        text:
          type: string
      required:
        - id
        - label
        - text
    QuestaoComResposta:
      type: object
      properties:
        id:
          type: string
        questionText:
          type: string
        mediaType:
          type: string
          enum: [sinal, foto, nenhum]
        imagePath:
          type: ["string", "null"]
        categoryId:
          type: string
        categoryName:
          type: string
        explanation:
          type: ["string", "null"]
        options:
          type: array
          items:
            $ref: "#/components/schemas/Opcao"
        correctOptionId:
          type: string
      required:
        - id
        - questionText
        - mediaType
        - imagePath
        - categoryId
        - categoryName
        - explanation
        - options
        - correctOptionId
...