# lib/api-zod/src/generated/types/opcao.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * RumoCarta API — exam prep platform for INATRO driving test
 * OpenAPI spec version: 0.1.0
 */

export interface Opcao {
  id: string;
  label: string;
  text: string;
}