# lib/api-zod/src/generated/types/questaoComResposta.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * RumoCarta API — exam prep platform for INATRO driving test
 * OpenAPI spec version: 0.1.0
 */
import type { Opcao } from './opcao';
import type { QuestaoComRespostaMediaType } from './questaoComRespostaMediaType';

export interface QuestaoComResposta {
  id: string;
  questionText: string;
  mediaType: QuestaoComRespostaMediaType;
  /** @nullable */
  imagePath: string | null;
  categoryId: string;
  categoryName: string;
  /** @nullable */
  explanation: string | null;
  options: Opcao[];
  correctOptionId: string;
}